Changed: #853 Compilation on 64-bits platforms

+++ code/ryzom/server/src/entities_game_service/building_manager/building_unit_test.cpp
@@ -670,7 +670,7 @@ NLMISC_COMMAND (testBuildingManager, "(debug) Unit test for building manager",
 		}
 	}
 
-	TestCharacters.resize( min(nbChars, FakeCharacters.size()) );
+	TestCharacters.resize( min(nbChars, (uint)FakeCharacters.size()) );
 	for (uint i = 0; i < TestCharacters.size(); i++)
 	{
 		CCharacter * c = PlayerManager.getChar( FakeCharacters[i] );
